requirements.txt
fastapi==0.121.1
uvicorn==0.38.0
SQLAlchemy==2.0.44
pydantic==2.12.2
pydantic-settings==2.0.3
pytest==7.4.0
httpx==0.24.1
requests==2.32.5
pymysql==1.1.0
cryptography==41.0.3
python-dotenv==1.0.0

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1

# Email and OTP dependencies
aiosmtplib==3.0.1
email-validator==2.1.0

# Image processing dependencies
Pillow==12.0.0